Split PostGIS installation and helper methods into separate files
