+++ backend/runners/initialize_postgres.py
@@ -1,8 +1,8 @@
 """Initialize Postgres tables."""
 from backend.lib.database.postgres import connect
-from backend.lib.database.postgres import postgis
 from backend.lib.database.postgres.base import Base
 from backend.lib.database.tables import address, provider, representative_point, service_area  # NOQA
+from backend.runners import install_postgis
 
 from sqlalchemy_utils import create_database, database_exists
 
@@ -15,9 +15,10 @@ def initialize_postgres():
     if not database_exists(postgres_engine.url):
         create_database(postgres_engine.url)
 
-    # Install Postgis
-    postgis.install()
-    # Create tables
+    # Install Postgis.
+    install_postgis.install()
+
+    # Create tables.
     Base.metadata.create_all(postgres_engine, checkfirst=True)
